# Fix sequential metadata fetching in ListingTable causing high latency