@geoffreyclaude 2025-02-28
> I pushed a commit to fix the formatting and try to get a clean CI run

Thanks! I had environment issues which prevented me from running the CI locally. Figured adding an extra basic unit test was simple enough to skip it 🤦