@alamb 2025-03-03
I tried to verify these changes but I couldn't figure out how to create an external table with explicitly listing the names via SQL.

For posterity here is what I tried:


This PR:
```
> create external table hits stored as parquet OPTIONS('aws.region' 'eu-central-1') location 's3://clickhouse-public-datasets/hits_compatible/athena_partitioned/';
0 row(s) fetched.
Elapsed 2.690 seconds.
```

Main
```
> create external table hits stored as parquet OPTIONS('aws.region' 'eu-central-1') location 's3://clickhouse-public-datasets/hits_compatible/athena_partitioned/';
0 row(s) fetched.
Elapsed 2.284 seconds.
```

I am going to try and whip up a smaller reproducer